src/imovel/requirements.txt
django
mysqlclient
djangorestframework
django-filter
drf-yasg
coreapi